% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{anyNA}
\alias{anyNA}
\title{Check for NAs}
\usage{
anyNA(x)
}
\arguments{
\item{x}{R object to be tested.}
}
\value{
A logical value. TRUE in case input has NA, FALSE otherwise.
}
\description{
During R update this function might be missing from the base package.
}
\examples{
no_NA <- 1:20
anyNA(no_NA)
present_NA <- rep(c(1:5, NA), 2)
anyNA(present_NA)

}
