% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Batten_disease_seeds}
\alias{Batten_disease_seeds}
\title{Batten disease seeds}
\format{An object of class \code{character} of length 44.}
\usage{
Batten_disease_seeds
}
\description{
Seed list of molecules associated with Batten disease
(juvenile neuronal ceroid lipofuscinosis).
The disease-associated molecules were retrieved from heterogeneous
databases such as NCBI Gene database [1], PolySearch [2],
DISEASES [3], Génie [4], Ingenuity Pathways Analysis [5],
PathwayStudio [6] and MetaCore [7]. All results were preprocessed
and merged into a seed list with unique entries (refer to [8]).
All molecules are anotated with HGNC gene symbols.
}
\references{
1. Maglott D, Pruitt K, Tatusova T, Murphy T. Gene.
In: The NCBI Handbook. 2nd ed. Bethesda (MD): National Center for
Biotechnology Information (US); 2013.

2. Cheng D, Knox C, Young N, Stothard P, Damaraju S,
Wishart DS. PolySearch: a web-based text mining system for extracting
relationships between human diseases, genes, mutations, drugs and
metabolites. Nucleic Acids Res. 2008;36:399–405.

3. Pletscher-Frankild S, Pallejà A, Tsafou K, Binder JX,
Jensen LJ. DISEASES: Text mining and data integration of disease-gene
associations. Methods. 2014;74:83–9.

4. Fontaine J-F, Priller F, Barbosa-Silva A,
Andrade-Navarro MA. Génie: literature-based gene prioritization at
multi genomic scale. Nucleic Acids Res. 2011;39:W455-61.

5. Racine J, Gerber V, Feutz MM, Riley CP, Adamec J,
Swinburne JE, et al. Comparison of genomic and proteomic data in
recurrent airway obstruction affected horses using Ingenuity Pathway
Analysis®. BMC Vet. Res. 2011;7:48.

6. Nikitin A, Egorov S, Daraselia N, Mazo I. Pathway
studio - The analysis and navigation of molecular networks.
Bioinformatics. 2003;19:2155–7.

7. Ekins S, Nikolsky Y, Bugrim A, Kirillov E, Nikolskaya T.
Pathway mapping tools for analysis of high content data. Methods Mol.
Biol. 2007;356:319–50.

8. Androsova G., Schneider R., Krause R. Study-driven
bias effects on disease-specific network extraction. (In preparation)
}
\keyword{datasets}
