% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{convert_to_adjacency_matrix}
\alias{convert_to_adjacency_matrix}
\title{Convert graph to adjacency matrix}
\usage{
convert_to_adjacency_matrix(graph, edge_attrib)
}
\arguments{
\item{graph}{An igraph graph with weights as edges attribute.}

\item{edge_attrib}{Name of the edge attribute with edges' weights.}
}
\value{
A symmetric matrix with number of rows and columns corresponding to the
number of vertices in the input graph.
}
\description{
This function was modified from \code{as_adjacency_matrix()} function from
'igraph' package to yield a faster computation for the large graphs. The graph
is represented as symmetric adjacency matrix, where rows and columns correspond
to graph vertices.
}
\note{
This function results in error if edge attribute with provided
\code{edge_attrib} name does not exist.
}
\examples{
graph <- igraph::make_tree(10, 4, mode = "undirected")
igraph::E(graph)$combined_score <- 1:igraph::ecount(graph)
convert_to_adjacency_matrix(graph, "combined_score")

}
\references{
Gábor Csárdi, Tamás Nepusz: The igraph software package for
complex network research. InterJournal Complex Systems, 1695, 2006.
}
