% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{degree_discounted_graph}
\alias{degree_discounted_graph}
\title{Degree-discounting of graph}
\usage{
degree_discounted_graph(graph, edge_attrib)
}
\arguments{
\item{graph}{An igraph graph with weights as edges attribute.}

\item{edge_attrib}{Name of the edge attribute with edges' weights.}
}
\value{
A degree-discounted graph. The edge weights in the degree-discounted
graph are normalized from 0 to 1.
}
\description{
This function reduces the weight of interactions (edges) proportionally
to the vertex degree. Degree-discounting was adapted from the notion of
reduced adjacency matrix in Laplacian matrix theory. This function requires
\code{graph} with weighted edges as an input.
}
\note{
This function results in error if edge attribute with provided
\code{edge_attrib} name does not exist.
}
\examples{
graph <- igraph::make_tree(10, 4, mode = "undirected")
E(graph)$combined_score <- 1:igraph::ecount(graph)
degree_discounted_graph(graph, "combined_score")

}
\references{
Satuluri V. and Parthasarathy S. (2011) Symmetrizations
for clustering directed graphs. In Proceedings of the 14th International
Conference on Extending Database Technology (EDBT/ICDT '11), ACM, New York, NY,
USA, 343-354.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Laplacian_matrix} for reduced
adjacency matrix in Laplacian matrix theory
}
