% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{degree_discounted_matrix}
\alias{degree_discounted_matrix}
\title{Degree-discounting of adjacency matrix}
\usage{
degree_discounted_matrix(adj_matrix, alpha = 0.5)
}
\arguments{
\item{adj_matrix}{A symmetric numeric matrix.}

\item{alpha}{A numeric value for the alpha parameter.}
}
\value{
A degree-discounted symmetric matrix with number of rows and columns
as in input adjacency matrix. The values in the degree-discounted matrix are
normalized from 0 to 1.
}
\description{
This function reduces the weight of interactions (edges) proportionally
to the vertex degree. Degree-discounting was adapted from the notion of
reduced adjacency matrix in Laplacian matrix theory. This function requires
\code{adj_matrix} and \code{alpha} parameter as an input. By default alpha
is set to -0.5 as suggested by Satuluri and Parthasarathy (2011).
}
\note{
This function results in error if the input adjacency matrix is
not symmetric.
}
\examples{
adjacency_matrix <- matrix(c(0, 0.85, 0.5, 0.5, 0, 0.9, 0.6,
0, 0, 0.85, rep(0,8), 0.5, rep(0,8), 0.5, 0, 0, 0, 0.5, 0,
0.8, 0.51, rep(0,4), 0.5, 0, 0.9, 0, 0, 0, 0.9, 0, 0, 0,
0.9, rep(0,4), 0.6, 0, 0, 0.8, rep(0,4), 0.6, 0, 0, 0,
0.51, rep(0,11), 0.6, 0, 0), ncol = 9)
colnames(adjacency_matrix) <- rownames(adjacency_matrix) <- LETTERS[1:9]
degree_discounted_matrix(adjacency_matrix)

}
\references{
Satuluri V. and Parthasarathy S. (2011) Symmetrizations
for clustering directed graphs. In Proceedings of the 14th International
Conference on Extending Database Technology (EDBT/ICDT '11), ACM, New York, NY,
USA, 343-354.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Laplacian_matrix} for more
information on reduced adjacency matrix.
}
