% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuting.R
\name{disnet_comm1}
\alias{disnet_comm1}
\title{disnet_comm1}
\usage{
disnet_comm1(j, edges_subset, i)
}
\arguments{
\item{j}{all the nodes `i` is connected to}

\item{edges_subset}{subset of `nd_edges` with all outgoing edges of `i`}

\item{i}{the node for which commuting proportion is being calculated}
}
\description{
gets commuting rate for each individual node (hence it's for individual nodes)
Considers a commuting proportion of 11% for the entire population
}
