% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuting.R
\name{disnet_comm2}
\alias{disnet_comm2}
\title{disnet_comm2}
\usage{
disnet_comm2(i, test_edges)
}
\arguments{
\item{i}{id of the `from` node, for which we want to calculate commuting
proportion}

\item{test_edges}{the dataframe of the edges which contains the distances
between the nodes and the population of each edges's}
}
\description{
Gets commuting rate for all nodes (includes individual node function)
}
