% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_commuting.R
\name{disnet_comm3}
\alias{disnet_comm3}
\title{disnet_comm3}
\usage{
disnet_comm3(g)
}
\arguments{
\item{g}{graph whose commuting rate needs to be calculated}
}
\description{
Preps the incoming graph for commuting rate calculation
}
