% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuting.R
\name{disnet_commuting}
\alias{disnet_commuting}
\title{Commuting function}
\usage{
disnet_commuting(g)
}
\arguments{
\item{g}{The `graphml` (network) object for which to calculate commuting rates}
}
\description{
`disnet_commuting` gets the `from` nodes, and sends them as the parameter for
the disnet_comm2 along with the entire dataframe `nd_edges`.
It incoporates both the individual and all nodes' functions.
Note: it removes nodes with less than 10 people in them.
}
\examples{
f = system.file("sampleData", "g.rds", package = "disnet")
g = readRDS(f)
disnet_commuting(g)
}
