% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-setup.R
\name{disnet_sim_setup}
\alias{disnet_sim_setup}
\title{Set up graph file to run simulations upon}
\usage{
disnet_sim_setup(g, r0 = 1.44, latent_period = 2.62, inf_period = 3.38,
  tau = 3, r_beta = 0.5, p_a = 1/3, seed_nd = NA, seed_no = 1,
  vacc = FALSE, output_dir = getOption("disnetOutputDir", NA),
  vacc_eff = c(80, 90), vacc_cov = c(80, 100), vacc_nd = "890", ...)
}
\arguments{
\item{g}{The network object created after commuting rates were calculated}

\item{r0}{Reproduction number}

\item{latent_period}{Period between infection and becoming infectious}

\item{inf_period}{Duration of infectious period}

\item{tau}{Return rate (from commuting)}

\item{r_beta}{Reduction in infectiousness due to being aymptomatic inf}

\item{p_a}{Probability of asymptomatic infection}

\item{seed_nd}{The node which you want to seed/initiate infection in}

\item{seed_no}{The number of infected people you want to start inf with}

\item{vacc}{Whether a vaccination campaign is underway or not}

\item{output_dir}{Directory to store intermediate simulation data in.}

\item{vacc_eff}{Vaccination efficacy}

\item{vacc_cov}{Vaccination coverage}

\item{vacc_nd}{Node you want to vaccinate in}
}
\description{
Set up graph file to run simulations upon
}
\details{
1. Reads in an `igraph` object (with commuting proportions previously
     calculated)

2. Calculates and attaches  "Effective Population" to graph

3. Vaccinates target nodes with vaccination proportion:
   vaccination proportions = vaccination coverage * vaccination efficacy
   The proportion of a node that is vaccinated is removed from the
   Susceptible compartment and added to the Recovered compartment

3. Seeds a node with specified number of infectious individuals (taking them
     out from the Susceptibles and adding to Infectious compartment

4. Creates the FOI structure/scaffold used to quickly calculate FOI
     (FOI = Force of Infection, acting on an individual in a node)
}
\examples{
f = system.file("sampleData", "g.rds", package = "disnet")
g = readRDS(f)
g_comm = disnet_commuting(g)
nodes = igraph::vcount(g_comm)
set.seed(890)
seed_nd = igraph::vertex_attr(g_comm, "name", sample(1:nodes, 1))
for_sim = disnet_sim_setup(g_comm, seed_nd = seed_nd, output_dir = NA)
}
