% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{first_neighborhood}
\alias{first_neighborhood}
\title{First-neighborhood graph}
\usage{
first_neighborhood(graph, seeds)
}
\arguments{
\item{graph}{An igraph graph with weights as edges attribute.}

\item{seeds}{A vector with unique values corresponding to the input
seed names.}
}
\value{
An igraph graph.
}
\description{
This function creates an igraph graph from the input \code{graph} and
set of \code{seeds}. It selects all directly connected nodes to the input
\code{seeds} disregarding weights of the edges.
}
\note{
This function results in error if the input vector \code{seeds}
has non-unique entries.
}
\examples{
library(STRINGdb)
string_db = STRINGdb$new(version="10", species=9606, score_threshold=500)
graph <- STRING_graph(string_db)
seeds_df <- data.frame(genes = Batten_disease_seeds)
maped_seeds <- string_db$map(seeds_df, "genes")
#Remove the species identifier from the protein names
seeds <- sapply(strsplit(maped_seeds[!is.na(maped_seeds[,2]),2], '\\\\.'), "[[", 2)
first_neighborhood(graph, seeds)

}
