% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_inf-info_sims.R
\name{get_inf_info}
\alias{get_inf_info}
\title{Compare outbreak information across simulations}
\usage{
get_inf_info(sims, sim_info, sim_l, threshold_nds = 100)
}
\arguments{
\item{sims}{Individual simulation results}

\item{sim_info}{Simulation information files}

\item{sim_l}{Simulation length information}
}
\description{
Compare outbreak information across simulations
}
