% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf-info_sims.R
\name{get_outbrks}
\alias{get_outbrks}
\title{Subset simulations that were outbreaks}
\usage{
get_outbrks(sims, inf_info)
}
\arguments{
\item{sims}{Simulation results}

\item{inf_info}{Infection information for each simulation}
}
\description{
Subset simulations that were outbreaks
}
