% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_nodes_to_adm.R
\name{get_outbrks_adm}
\alias{get_outbrks_adm}
\title{Join simulation nodes with WHO adm regions}
\usage{
get_outbrks_adm(outbrk_TS, net_pts_join)
}
\arguments{
\item{outbrk_TS}{Time-step dataframes for each simulation}

\item{net_pts_join}{Network point dataframe containing country and WHO information
mapped to each node name.}
}
\description{
Join simulation nodes with WHO adm regions
}
