% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_to_WHO.R
\name{get_outbrks_ctry}
\alias{get_outbrks_ctry}
\title{Join simulation nodes with countries}
\usage{
get_outbrks_ctry(outbrk_TS, net_pts_join)
}
\arguments{
\item{outbrk_TS}{Time-step dataframes for each simulation}

\item{net_pts_join}{Network point dataframe containing country and WHO information
mapped to each node name.}
}
\description{
Join simulation nodes with countries
}
