% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_inf_info.R
\name{get_outbrks_info}
\alias{get_outbrks_info}
\title{Subset information for outbreaks}
\usage{
get_outbrks_info(n, sim_info, inf_info)
}
\arguments{
\item{n}{Iteration}

\item{sim_info}{Simulation information}

\item{inf_info}{Infection information for each simulation}
}
\description{
Subset information for outbreaks
}
