% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_read-sims.R
\name{get_sim_info}
\alias{get_sim_info}
\title{Get simulation information function}
\usage{
get_sim_info(sim_dir, nsims = NA)
}
\arguments{
\item{sim_dir}{The directory containing simulation results.}
}
\description{
Reads in the simulation information files
}
