% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_read-sims.R
\name{get_sim_l}
\alias{get_sim_l}
\title{Get simulation durations/lengths}
\usage{
get_sim_l(sim_info_df)
}
\arguments{
\item{sim_info_df}{The simulation information files for each simulation}
}
\description{
This function uses the simulation information files to get the lengths of
each simulations
}
