% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_plot.R
\name{get_sim_lines}
\alias{get_sim_lines}
\title{Get simulation lines for plotting}
\usage{
get_sim_lines(i, f_outbrks_info, f_outbrks_l)
}
\arguments{
\item{i}{iteration over simulations to plot}

\item{f_outbrks_info}{location of file with outbreak information}

\item{f_outbrks_l}{location of file with outbreak lengths information}
}
\description{
Get simulation lines for plotting
}
