% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf-info_sims.R
\name{get_sim_summ}
\alias{get_sim_summ}
\title{Get summary of outbreak types}
\usage{
get_sim_summ(inf_info)
}
\arguments{
\item{inf_info}{Information on outbreak infections wrt nodes affected,
outbreak lengths, and numbers infected.}
}
\description{
Get summary of outbreak types
}
