% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sims.R
\name{get_sims}
\alias{get_sims}
\title{Read in individual simulation files}
\usage{
get_sims(sim_dir)
}
\arguments{
\item{sim_dir}{The directory containing simulation results.}
}
\description{
Imports each individual simulation file for the set of simulations.
}
