% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{greedy_search}
\alias{greedy_search}
\title{Greedy search graph}
\usage{
greedy_search(graph, seeds, steps, attrib_name)
}
\arguments{
\item{graph}{An igraph graph with weights as edges attribute.}

\item{seeds}{A vector with unique values corresponding to the input
seed names.}

\item{steps}{An integer number indicating number of steps of the greedy search}

\item{attrib_name}{Name of the edge attribute with edges' weights.}
}
\value{
An igraph graph.
}
\description{
This function creates an igraph graph from the input \code{graph} and
set of \code{seeds}. Each step it selects the strongest interacting nodes
defined by the edge attribute \code{attrib_name}. It makes number of
steps in the graph as defined by \code{steps} and retrieves unique
nodes not yet added in the previous step.
}
\note{
This function results in error if the input vector \code{seeds}
has non-unique entries or if edge attribute with provided
\code{edge_attrib} name does not exist.
}
\examples{
library(STRINGdb)
string_db = STRINGdb$new(version="10", species=9606, score_threshold=500)
graph <- STRING_graph(string_db)
seeds_df <- data.frame(genes = Batten_disease_seeds)
maped_seeds <- string_db$map(seeds_df, "genes")
#Remove the species identifier from the protein names
seeds <- sapply(strsplit(maped_seeds[!is.na(maped_seeds[,2]),2], '\\\\.'), "[[", 2)
greedy_search(graph, seeds, 3, "combined_score")

}
