% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_net.R
\name{map_g}
\alias{map_g}
\title{Function to map the network}
\usage{
map_g(g, nodes = NULL)
}
\arguments{
\item{g}{The network to be mapped.}

\item{nodes}{The nodes to be highlighted (in red vs. the default blue).}
}
\description{
This function creates a leaflet map of the network, and highlights nodes
of interest specified by the user.
}
\examples{
f = system.file("sample_data/network", "g100.RDS", package = "SEEDNet")
g = readRDS(f)
map = map_g(g)
}
