% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{observed_pred_table}
\alias{observed_pred_table}
\title{Table with observed-predicted proteins for permutation setup}
\usage{
observed_pred_table(graph_vertices, seeds)
}
\arguments{
\item{graph_vertices}{A vector with unique values corresponding to the input
graph vertex names.}

\item{seeds}{A vector with unique values corresponding to the input
seed names.}
}
\value{
A data.frame with graph vertices in rows and two columns for present
seeds and predicted during permutations.
}
\description{
This function creates a new table with all vertices from the input graph
and sets value to 1 for the input \code{seeds} in the "true" column,
while second column with be used for permutations.
}
\note{
This function results in error if the input vectors have non-unique entries.
}
\examples{
library(STRINGdb)
string_db = STRINGdb$new(version="10", species=9606, score_threshold=500)
graph <- STRING_graph(string_db)
seeds_df <- data.frame(genes = Batten_disease_seeds)
maped_seeds <- string_db$map(seeds_df, "genes")
#Remove the species identifier from the protein names
seeds <- sapply(strsplit(maped_seeds[!is.na(maped_seeds[,2]),2], '\\\\.'), "[[", 2)
graph_vertices <- igraph::V(graph)$name
observed_pred <- observed_pred_table(graph_vertices, seeds)
table(observed_pred)

}
