% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{shortest_path}
\alias{shortest_path}
\title{Shortest path graph}
\usage{
shortest_path(graph, seeds, edge_weights)
}
\arguments{
\item{graph}{An igraph graph with weights as edges attribute.}

\item{seeds}{A vector with unique values corresponding to the input
seed names.}

\item{edge_weights}{A numeric vector with edge weights (optional, see documentation
on \code{igraph::get.shortest.paths()}).}
}
\value{
An igraph graph.
}
\description{
This function is a wraper for igraph function \code{get.shortest.paths()}. It creates
igraph graph from the input \code{graph} and specified set of vertices \code{seeds}.
}
\note{
This function results in error if the input vector \code{seeds}
has non-unique entries.
}
\examples{
library(STRINGdb)
string_db = STRINGdb$new(version="10", species=9606, score_threshold=500)
graph <- STRING_graph(string_db)
seeds_df <- data.frame(genes = Batten_disease_seeds)
maped_seeds <- string_db$map(seeds_df, "genes")
# Remove the species identifier from the protein names
seeds <- sapply(strsplit(maped_seeds[!is.na(maped_seeds[,2]),2], '\\\\.'), "[[", 2)
#In STRING database, the higher combined score indicates the higher evidence (probability)
# of interaction between proteins, thus for shortest path (which selects the minimal
# distance between vertices based on the smallest sum of edges weights) we use reversed
# weights of the STRING graph.
shortest_path(graph, seeds, 1000-E(graph)$combined_score)

}
