% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seednet_functions.R
\name{STRING_graph}
\alias{STRING_graph}
\title{Load STRING protein-protein interactions}
\usage{
STRING_graph(string_db)
}
\arguments{
\item{string_db}{STRINGdb reference class.}
}
\value{
igraph object of connected graph from STRING database.
}
\description{
This function creates an igraph object from STRING database \code{string_db}.
It removes species identifier from the proteins names and returns the
largest connected component. Note: you must have an Internet connection to
get a STRINGdb reference class.
}
\examples{
library(STRINGdb)
string_db = STRINGdb$new(version="10", species=9606, score_threshold=500)
graph <- STRING_graph(string_db)
print(graph)

}
\references{
Franceschini, A et al. (2013). STRING v9.1: protein-protein
interaction networks, with increased coverage and integration. In:’Nucleic Acids Res.
2013 Jan;41(Database issue):D808-15. doi: 10.1093/nar/gks1094. Epub 2012 Nov 29
}
