% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_commuting.R
\name{disnet_comm2}
\alias{disnet_comm2}
\title{disnet_comm2}
\usage{
disnet_comm2(i, test_edges, N_c)
}
\arguments{
\item{i}{id of the `from` node, for which we want to calculate commuting
proportion}

\item{test_edges}{the dataframe of the edges which contains the distances
between the nodes and the population of each edges's}

\item{N_c}{The commuting proportion of the entire population. Default = 0.11 from
Simini}
}
\description{
Gets commuting rate for all nodes (includes individual node function)
}
