% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_inf_info.R
\name{get_inf_info}
\alias{get_inf_info}
\title{Compare outbreak information across simulations}
\usage{
get_inf_info(n, sims, sim_info, sim_l)
}
\arguments{
\item{n}{The iteration number}

\item{sims}{Individual simulation results}

\item{sim_info}{Simulation information files}

\item{sim_l}{Simulation length information}
}
\description{
Compare outbreak information across simulations
}
